Added: ProceduralAudioSource.cs ProceduralAudioSourceEditor.cs - Custom inspector for ProceduralAudioSource. Allows user (game designer/whatever) to easily change settings of the audio like frequency Added PublicAudioUtil.cs - Allows us to use functions inside AudioUtils.cs (part of unity's API) that we normally can't use. Added SaveWavUtil.cs - Lets us save created audiosource.
Todo:
Fix issues with playing the audio source. Unity has a method for playing audio in the editor inside AudioUtils.cs called PlayClip and we have to use reflection to be able to use it. It seems that the function cannot play AudioClips that haven't been saved so I'm trying to use SaveWavUtil to achieve this. Currently it just saves the audio each frame (Inside OnSceneGUI) and you can probably guess why that's a disgusting idea. Maybe I should only let the user play it once they saved it (via a button)